requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
PyMySQL==1.0.2
pyrsistent==0.19.3
python-decouple==3.7
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
